+++ tools/merge_tweets_csv.py
@@ -66,7 +66,7 @@ def merge_csv(input_folder, output_file, date_column):
 if __name__ == "__main__":
     # Folder tempat menyimpan semua file CSV yang mau digabung
     merge_csv(
-        input_folder="tools/tweets",        # Folder input
-        output_file="tools/tweets/merged.csv",  # Nama file hasil gabungan
-        date_column="created_at"            # Kolom tanggal untuk sortir (saat ini 'created_at')
+        input_folder="result/mbg/tweets",            # Folder input
+        output_file="result/mbg/tweets/merged.csv",  # Nama file hasil gabungan
+        date_column="created_at"                     # Kolom tanggal untuk sortir (saat ini 'created_at')
     )